Refactor some styles, add Image component
