+++ src/components/appHeader/appHeader.module.scss
@@ -45,3 +45,7 @@
 		}
 	}
 }
+
+.active {
+	color: $main-color;
+}
