+++ src/components/charItem/CharItem.tsx
@@ -1,6 +1,6 @@
 import cx from "classnames";
 
-import { IMGNOTFND } from "constant";
+import { Image } from "components/common/image";
 
 import { ITransformedChar } from "types";
 
@@ -15,7 +15,7 @@ interface ICharItemProps {
 const CharItem = ({ char: { name, thumbnail }, isActive, onClick }: ICharItemProps) => {
 	return (
 		<li className={cx(s.root, isActive && s.active)} tabIndex={0} onClick={onClick}>
-			<img src={thumbnail} alt={name} style={{ objectFit: thumbnail === IMGNOTFND ? "contain" : "cover" }} />
+			<Image src={thumbnail} altText={name} />
 			<div className={s.text}>{name}</div>
 		</li>
 	);
