+++ src/components/common/image/index.ts
@@ -1,0 +1,1 @@
+export * from "./Image";
